Moved remaining shared components
