+++ src/shared/components/default-error/prop-types.js
@@ -1,0 +1,11 @@
+import PropTypes from 'prop-types'
+
+export const defaultProps = {
+  message: 'Application Error'
+}
+
+export default {
+  message: PropTypes.string,
+  code: PropTypes.number,
+  children: PropTypes.node
+}
